pix2pix/requirements.txt
torch
torchvision
Pillow
scikit-learn
gdown
matplotlib